Add Java files, pom.xml, and some elements of .gitignore
This project will now include the very basic things needed to get started with a Java project. The current Java version target is 14, and it will probably make sense to change to the most recent LTS release of Java before the class begins.

My goal here was to get the minimum needed to run something as a Java project in VS Code so I could begin exploring using that environment for CSci 2101.